responsible_ai_automation/requirements.txt
numpy>=1.24.0
torch>=2.0.0
gymnasium>=0.29.0
stable-baselines3>=2.0.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0
tensorboard>=2.13.0
fairlearn>=0.9.0
shap>=0.42.0
aif360>=0.5.0
python-dotenv>=1.0.0
schedule>=1.2.0
psutil>=5.9.0

